feat: migrate metadata_shard binary format v2→v3 (u32→u64 fields)
- Bump MDB_SHARD_HEADER_VERSION from 2 to 3
- Bump MDB_SHARD_FOOTER_VERSION from 1 to 2
- Change all numeric metadata_shard fields from u32 to u64:
  FileDataSequenceHeader.num_entries
  FileDataSequenceEntry.{unpacked_segment_bytes, chunk_index_start, chunk_index_end}
  XorbChunkSequenceHeader.{num_entries, num_bytes_in_xorb, num_bytes_on_disk}
  XorbChunkSequenceEntry.{chunk_byte_range_start, unpacked_segment_bytes}
- Uniform 64-byte entry size (was 48 bytes)
- Pass version param through deserialization chain for future v2 compat
- Remove as u64 widening casts throughout
- Update callers and tests across xet_core, xet_adapter, fsck, rebuild, fuzz, server
- 7,963 tests pass

+++ crates/fsck/src/tests.rs
@@ -1323,12 +1323,12 @@ async fn run_fsck_with_invalid_dedupe_shard_mapping_detected() {
     let mut shard = MDBInMemoryShard::default();
     shard
         .add_file_reconstruction_info(MDBFileInfo {
-            metadata: FileDataSequenceHeader::new(file_hash_val, 1u32, false, false),
+            metadata: FileDataSequenceHeader::new(file_hash_val, 1u64, false, false),
             segments: vec![FileDataSequenceEntry::new(
                 xorb_val,
-                chunk_data.len() as u32,
-                0_u32,
-                1_u32,
+                chunk_data.len() as u64,
+                0_u64,
+                1_u64,
             )],
             verification: Vec::new(),
             metadata_ext: None,
@@ -1336,9 +1336,9 @@ async fn run_fsck_with_invalid_dedupe_shard_mapping_detected() {
         .unwrap();
     shard
         .add_xorb_block(MDBXorbInfo {
-            metadata: XorbChunkSequenceHeader::new(xorb_val, 1_u32, chunk_data.len() as u32),
+            metadata: XorbChunkSequenceHeader::new(xorb_val, 1_u64, chunk_data.len() as u64),
             chunks: vec![
-                XorbChunkSequenceEntry::new(chunk_hash, chunk_data.len() as u32, 0_u32)
+                XorbChunkSequenceEntry::new(chunk_hash, chunk_data.len() as u64, 0_u64)
                     .with_global_dedup_flag(true),
             ],
         })
@@ -1422,12 +1422,12 @@ async fn run_fsck_with_dedupe_shard_added_to_reachability() {
     let mut shard = MDBInMemoryShard::default();
     shard
         .add_file_reconstruction_info(MDBFileInfo {
-            metadata: FileDataSequenceHeader::new(file_hash_val, 1u32, false, false),
+            metadata: FileDataSequenceHeader::new(file_hash_val, 1u64, false, false),
             segments: vec![FileDataSequenceEntry::new(
                 xorb_val,
-                chunk_data.len() as u32,
-                0_u32,
-                1_u32,
+                chunk_data.len() as u64,
+                0_u64,
+                1_u64,
             )],
             verification: Vec::new(),
             metadata_ext: None,
@@ -1435,9 +1435,9 @@ async fn run_fsck_with_dedupe_shard_added_to_reachability() {
         .unwrap();
     shard
         .add_xorb_block(MDBXorbInfo {
-            metadata: XorbChunkSequenceHeader::new(xorb_val, 1_u32, chunk_data.len() as u32),
+            metadata: XorbChunkSequenceHeader::new(xorb_val, 1_u64, chunk_data.len() as u64),
             chunks: vec![
-                XorbChunkSequenceEntry::new(chunk_hash, chunk_data.len() as u32, 0_u32)
+                XorbChunkSequenceEntry::new(chunk_hash, chunk_data.len() as u64, 0_u64)
                     .with_global_dedup_flag(true),
             ],
         })
